@@ -1,138 +1,69 @@
 package org.deepspeechdemo
 
-import android.Manifest
-import android.content.pm.PackageManager
-import android.media.AudioFormat
-import android.media.AudioRecord
-import android.media.MediaRecorder
-import android.os.Build
 import android.os.Bundle
-import android.view.View
 import androidx.appcompat.app.AppCompatActivity
-import androidx.core.app.ActivityCompat
-import kotlinx.android.synthetic.main.activity_main.*
-import org.mozilla.deepspeech.libdeepspeech.DeepSpeechModel
-import java.io.File
-import java.util.concurrent.atomic.AtomicBoolean
+import androidx.databinding.DataBindingUtil
+import androidx.lifecycle.ViewModelProvider
+import androidx.lifecycle.get
+import androidx.lifecycle.viewModelScope
+import androidx.lifecycle.whenResumed
+import kotlinx.coroutines.ObsoleteCoroutinesApi
+import kotlinx.coroutines.launch
+import org.deepspeechdemo.databinding.ActivityMainBinding
 
+@ObsoleteCoroutinesApi
 class MainActivity : AppCompatActivity() {
-    private var model: DeepSpeechModel? = null
-
-    private var transcriptionThread: Thread? = null
-    private var isRecording: AtomicBoolean = AtomicBoolean(false)
-
-    private val TFLITE_MODEL_FILENAME = "deepspeech-0.8.0-models.tflite"
-    private val SCORER_FILENAME = "deepspeech-0.8.0-models.scorer"
-
-    private fun checkAudioPermission() {
-        // Permission is automatically granted on SDK < 23 upon installation.
-        if (Build.VERSION.SDK_INT >= 23) {
-            val permission = Manifest.permission.RECORD_AUDIO
-
-            if (checkSelfPermission(permission) != PackageManager.PERMISSION_GRANTED) {
-                ActivityCompat.requestPermissions(this, arrayOf(permission), 3)
-            }
-        }
-    }
-
-    private fun transcribe() {
-        // We read from the recorder in chunks of 2048 shorts. With a model that expects its input
-        // at 16000Hz, this corresponds to 2048/16000 = 0.128s or 128ms.
-        val audioBufferSize = 2048
-        val audioData = ShortArray(audioBufferSize)
-
-        runOnUiThread { btnStartInference.text = "Stop Recording" }
-
-        model?.let { model ->
-            val streamContext = model.createStream()
-
-            val recorder = AudioRecord(
-                MediaRecorder.AudioSource.VOICE_RECOGNITION,
-                model.sampleRate(),
-                AudioFormat.CHANNEL_IN_MONO,
-                AudioFormat.ENCODING_PCM_16BIT,
-                audioBufferSize
-            )
-            recorder.startRecording()
-
-            while (isRecording.get()) {
-                recorder.read(audioData, 0, audioBufferSize)
-                model.feedAudioContent(streamContext, audioData, audioData.size)
-                val decoded = model.intermediateDecode(streamContext)
-                runOnUiThread { transcription.text = decoded }
-            }
-
-            val decoded = model.finishStream(streamContext)
-
-            runOnUiThread {
-                btnStartInference.text = "Start Recording"
-                transcription.text = decoded
-            }
-
-            recorder.stop()
-            recorder.release()
-        }
-    }
-
-    private fun createModel(): Boolean {
-        val modelsPath = getExternalFilesDir(null).toString()
-        val tfliteModelPath = "$modelsPath/$TFLITE_MODEL_FILENAME"
-        val scorerPath = "$modelsPath/$SCORER_FILENAME"
-
-        for (path in listOf(tfliteModelPath, scorerPath)) {
-            if (!File(path).exists()) {
-                status.append("Model creation failed: $path does not exist.\n")
-                return false
-            }
-        }
-
-        model = DeepSpeechModel(tfliteModelPath)
-        model?.enableExternalScorer(scorerPath)
-
-        return true
-    }
-
-    private fun startListening() {
-        if (isRecording.compareAndSet(false, true)) {
-            transcriptionThread = Thread(Runnable { transcribe() }, "Transcription Thread")
-            transcriptionThread?.start()
-        }
-    }
+    private var core: DSCore? = null
+    private lateinit var vm: DSViewModel
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
-        setContentView(R.layout.activity_main)
-        checkAudioPermission()
-
-        // Create application data directory on the device
-        val modelsPath = getExternalFilesDir(null).toString()
-
-        status.text = "Ready. Copy model files to \"$modelsPath\" if running for the first time.\n"
-    }
-
-    private fun stopListening() {
-        isRecording.set(false)
-    }
-
-    fun onRecordClick(v: View?) {
-        if (model == null) {
-            if (!createModel()) {
-                return
+        DataBindingUtil.setContentView<ActivityMainBinding>(this, R.layout.activity_main).run {
+            vm = ViewModelProvider(this@MainActivity).get<DSViewModel>().apply {
+                recordButton.setOnClickListener {
+                    core!!.startStopTranscription()
+                }
+                transcriptionMut.value = getString(R.string.spoken_text_will_appear_here)
+                (application as DSApp).addModelReloadListener(this@MainActivity) {
+                    core?.run {
+                        viewModelScope.launch {
+                            whenResumed {
+                                stopTranscription()
+                                close()
+                                core = null
+                                loadModel(this@MainActivity)?.use {
+                                    core = DSCore(this@apply, it)
+                                }
+                            }
+                        }
+                    }
+                }
+                this@MainActivity.vm = this
             }
-            status.append("Created model.\n")
+            lifecycleOwner = this@MainActivity
         }
+    }
 
-        if (isRecording.get()) {
-            stopListening()
-        } else {
-            startListening()
+    override fun onResume() {
+        super.onResume()
+        if (!checkAudioPermission(this)) {
+            return
         }
+        if (core == null) {
+            core = vm.setupCoreOrPrompt(this, {
+                startStopTranscription()
+            })
+        }
+    }
+
+    override fun onPause() {
+        super.onPause()
+        core?.stopTranscription()
     }
 
     override fun onDestroy() {
         super.onDestroy()
-        if (model != null) {
-            model?.freeModel()
-        }
+        core?.close()
+        core = null
     }
 }
